# specs/017-better-auth-integration/contracts/auth-api-contract.yaml
openapi: 3.0.3
info:
  title: Better Auth Integration API
  description: API contract for authentication and user isolation
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  minLength: 8
                  pattern: "^(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*]).*$"
                  description: User's password with minimum requirements
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                    format: uuid
        '400':
          description: Invalid input or user already exists
  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticates user and returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  refresh_token:
                    type: string
                    description: JWT refresh token
                  token_type:
                    type: string
                    enum: [bearer]
                  user_id:
                    type: string
                    format: uuid
        '401':
          description: Invalid credentials
...